Address some comments.
